Successfully saved csv file to a .db file using sqlite3
A piece of functionality that could be useful is trimming the csv file itself from this file, though the practical limits for what is considered a reasonable number of records in a csv file has not been established in this project yet. Therefore, it would be difficult to know how many records to trim each time. The seperation could be a logical one, where only non-stale records go into the csv, and stale records get archived regularly to the .db file. This could make accessing the stale records more efficient, such as when we want to do data access in sql, rather than trying to open a massive csv file in excel which has practical limits of about 1 million unique records. So the reason for putting things into the sqlite3 db file is to make later data retrieval easier.